Chap 4 terminè

+++ controllers/stuff.js
@@ -1,4 +1,5 @@
 const Thing = require('../models/Thing');
+const fs = require('fs');
 
 exports.createThing = (req, res, next) => {
     const thingObject = JSON.parse(req.body.thing);
@@ -15,15 +16,45 @@ exports.createThing = (req, res, next) => {
 };
 
 exports.modifyThing = (req, res, next) => {
-    Thing.updateOne({_id: req.params.id}, {...req.body, _id: req.params.id})
-        .then((thing) =>{ res.status(200).json({message: 'Objet modifié'})})
-        .catch((error) => {res.status(400).json({error: error})});
+    const thingObject = req.file ? {            // Si un nouveau fichier est téléchargé
+        ...JSON.parse(req.body.thing),
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    } : {
+        ...req.body
+    };
+
+    delete thingObject._userId;
+    Thing.findOne({_id: req.params.id})
+        .then((thing) =>{
+            if(thing.userId !== req.auth.userId){
+                res.status(401).json({message: "Non-autorisé"});
+            }
+            else{
+                Thing.updateOne({_id: req.params.id}, {...thingObject, _id: req.params.id})
+                    .then(() =>{ res.status(200).json({message: 'Objet modifié'})})
+                    .catch((error) => {res.status(401).json({error})});
+            }
+        })
+        .catch((error) => {res.status(400).json({error})});
 };
 
 exports.deleteThing = (req, res, next) => {
-    Thing.deleteOne({_id: req.params.id})
-        .then((thing) => {res.status(200).json({message: 'Objet supprimé'})})
-        .catch((error) => {res.status(400).json({error: error})});
+    Thing.findOne({_id: req.params.id})
+        .then((thing) => {
+            if(thing.userId !== req.auth.userId){
+                res.status(401).json({message: "Non-autorisé"});
+            }
+            else{
+                const filename = thing.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Thing.deleteOne({_id: req.params.id})
+                        .then(() => {res.status(200).json({message: 'Objet supprimé'})})
+                        .catch((error) => {res.status(400).json({error: error})});  
+                });
+            } 
+            
+        })
+        .catch((error) => {res.status(404).json({error})});
 };
 
 exports.getOneThing = (req, res, next) => {
